requirements.txt
# Python Backend Dependencies for ChatBot RAG System

# Core Environment Management
python-dotenv

# Google AI & LLM Integration
google-generativeai
langchain
langchain-core
langchain-google-genai

# Document Processing & Embeddings
langchain-community
langchain-chroma
chromadb
PyPDF2

# Semantic Processing
semantic-router
semantic-chunkers

# Utilities
pydantic
python-multipart

# Web Framework
fastapi
uvicorn
